FINALLY got the kinks worked out of backpropagation.
While this is still in a very rough state, I've chosen to commit
what I have and clean it up in later commits.

Backpropagation works well enough to approximate a sine wave
for two periods with an average error of 0.06, and it's fairly
fast, at least when you send all the debugging output to a file.
Getting to this stage was an accomplishment, and basically, I
managed to fix three main issues with the networks:

  - The derivative for the impulse functions was set up to
  take a value for x when it was really being passed a value
  of f(x). One of the biggest results of fixing this was a
  dramatic speed up in backpropagation performance.
  - The network was using weights that had already been modified
  by the packpropagation algorithm to calculate the error
  for the next layer.
  - The network machinery was using a sigmoid activation function
  for the input neurons, instead of no activation function like
  it should be. The backpropagation algorithm, however, asumed
  that is was not using an activation function, and therefore,
  neglected to touch the weights of the input neurons, causing
  the overall error to converge upon weird values.